tests: move tests for db.utils.clear_my_address
